requirements.txt
dotenv
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib

# iki treciadienio google callender
# ka reiskia kad php veikia?
